+++ game_patch/os/frametime.cpp
@@ -12,6 +12,7 @@
 #include "../rf/os/frametime.h"
 #include "../multi/multi.h"
 #include "../main/main.h"
+#include "../misc/alpine_settings.h"
 #include "../hud/hud.h"
 #include <xlog/xlog.h>
 
@@ -46,12 +47,12 @@ static void frametime_render_fps_counter()
     if (g_game_config.fps_counter && !rf::hud_disabled) {
         auto text = std::format("FPS: {:.1f}", rf::current_fps);
         rf::gr::set_color(0, 255, 0, 255);
-        int x = rf::gr::screen_width() - (g_game_config.big_hud ? 165 : 90);
+        int x = rf::gr::screen_width() - (g_alpine_game_config.big_hud ? 165 : 90);
         int y = 10;
         if (rf::gameseq_in_gameplay()) {
-            y = g_game_config.big_hud ? 110 : 60;
+            y = g_alpine_game_config.big_hud ? 110 : 60;
             if (hud_weapons_is_double_ammo()) {
-                y += g_game_config.big_hud ? 80 : 40;
+                y += g_alpine_game_config.big_hud ? 80 : 40;
             }
         }
 
@@ -62,12 +63,12 @@ static void frametime_render_fps_counter()
     if (g_game_config.ping_display && !rf::hud_disabled && rf::is_multi && !rf::is_server) {
         auto text = std::format("Ping: {}", rf::local_player->net_data->ping);
         rf::gr::set_color(0, 255, 0, 255);
-        int x = rf::gr::screen_width() - (g_game_config.big_hud ? 165 : 90);
-        int y = g_game_config.big_hud ? 35 : 25;
+        int x = rf::gr::screen_width() - (g_alpine_game_config.big_hud ? 165 : 90);
+        int y = g_alpine_game_config.big_hud ? 35 : 25;
         if (rf::gameseq_in_gameplay()) {
-            y = g_game_config.big_hud ? 135 : 75;
+            y = g_alpine_game_config.big_hud ? 135 : 75;
             if (hud_weapons_is_double_ammo()) {
-                y += g_game_config.big_hud ? 105 : 55;
+                y += g_alpine_game_config.big_hud ? 105 : 55;
             }
         }