+++ game_patch/rf/input.h
@@ -119,6 +119,7 @@ namespace rf
 
     static auto& mouse_get_pos = addr_as_ref<int(int &x, int &y, int &z)>(0x0051E450);
     static auto& mouse_was_button_pressed = addr_as_ref<int(int btn_idx)>(0x0051E5D0);
+    static auto& mouse_button_is_down = addr_as_ref<bool(int btn_idx)>(0x0051E530);
     static auto& mouse_set_visible = addr_as_ref<void(bool visible)>(0x0051E680);
     static auto& key_process_event = addr_as_ref<void(int scan_code, int key_down, int delta_time)>(0x0051E6C0);
     static auto& key_is_down = addr_as_ref<bool(Key key)>(0x0051F220);